refactor: UpdateExpenseCategoryModal 리팩토링
