+++ src/app/home/setting/expenseCategory/_components/UpdateExpenseCategoryModal.tsx
@@ -3,66 +3,22 @@ import NagativeButton from "@/components/button/NagativeButton";
 import ModalForm from "@/components/Modal/ModalForm";
 import ModalName from "@/components/Modal/ModalName";
 import ModalWrapper from "@/components/Modal/ModalWrapper";
-import useModalStore from "@/stores/modalStore";
-import { useEffect, useRef } from "react";
-import { ExpenseCategoryAction } from "../page";
-import { updateExpenseCategory } from "@/apis/services/expenseCategory";
-
-interface UpdateExpenseCategoryModalProps {
-    currentExpenseCategory: {
-        id: number;
-        name: string;
-    };
-    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
-}
+import useUpdateExpenseCategoryModal from "./UpdateExpenseCategoryModal.hook";
+import { UpdateExpenseCategoryModalProps } from "./UpdateExpenseCategoryModal.type";
 
 export default function UpdateExpenseCategoryModal({
     currentExpenseCategory,
     expenseCategoriesDispatch,
 }: UpdateExpenseCategoryModalProps) {
-    const { closeModal } = useModalStore();
-
-    const inputRef = useRef<HTMLInputElement>(null);
-
-    const handleUpdateButtonClick = () => {
-        // 낙관적 업데이트
-        if (!inputRef.current) return;
-
-        expenseCategoriesDispatch({
-            type: "UPDATE",
-            payload: {
-                id: currentExpenseCategory.id,
-                name: inputRef.current?.value,
-            },
-        });
-
-        // api 요청
-        const formValues = {
-            id: currentExpenseCategory.id,
-            name: inputRef.current.value,
-        };
-
-        updateExpenseCategory({ formValues: formValues }).catch((error) => {
-            console.error(error);
-        });
-
-        closeModal();
-    };
-
-    const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        if (!inputRef.current) return;
-        inputRef.current.value = e.target.value;
-    };
-
-    const handleModalCloseButtonClick = () => {
-        closeModal();
-    };
-
-    useEffect(() => {
-        if (currentExpenseCategory && inputRef.current) {
-            inputRef.current.value = currentExpenseCategory.name;
-        }
-    }, [currentExpenseCategory]);
+    const {
+        handleModalCloseButtonClick,
+        handleInputChange,
+        handleUpdateButtonClick,
+        inputRef,
+    } = useUpdateExpenseCategoryModal({
+        currentExpenseCategory,
+        expenseCategoriesDispatch,
+    });
 
     return (
         <ModalWrapper>
