+++ src/app/home/setting/expenseCategory/_components/UpdateExpenseCategoryModal.type.ts
@@ -1,0 +1,17 @@
+import { ExpenseCategoryAction } from "../page";
+
+export interface UpdateExpenseCategoryModalProps {
+    currentExpenseCategory: {
+        id: number;
+        name: string;
+    };
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
+
+export interface UseUpdateExpenseCategoryModalProps {
+    currentExpenseCategory: {
+        id: number;
+        name: string;
+    };
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
